services/vessel-memory/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Database drivers
psycopg2-binary==2.9.9
cassandra-driver==3.28.0

# Vector store
qdrant-client==1.7.0

# Knowledge graph
neo4j==5.14.1

# Embeddings
openai==1.3.7

# Kafka
kafka-python==2.0.2

# HTTP client
httpx==0.25.2

# Utilities
python-dotenv==1.0.0

# Made with Bob